Basic Level Progression
once the timer hits 10 seconds the word scrambles and a enemy with spikes walks across the screen that the player has to jump over, this scales currently up to two enemies,  and then the word continues to scramble  every 10 seconds until solved. once solved it wipes the board and clears the enemies off. I still need to add a player health system